refactor: 필터 관련 store 위치 수정
[#382]
